+++ src/extension.ts
@@ -2,6 +2,7 @@ import * as vscode from 'vscode';
 import { registerPlayDialogueAudioCommand } from './commands/playDialogueAudio';
 import { registerStopAudioCommand } from './commands/stopAudio';
 import { registerFixSilesianCommand } from './commands/fixSilesian';
+import { registerVerifyDialoguesCommand } from './commands/verifyDialogues';
 import { registerCodeLensProvider } from './providers/codeLensProvider';
 import { checkAndPromptForAudioDirectory } from './utils/configCheck';
 
@@ -15,6 +16,7 @@ export function activate(context: vscode.ExtensionContext) {
 	registerPlayDialogueAudioCommand(context);
 	registerStopAudioCommand(context);
 	registerFixSilesianCommand(context);
+	registerVerifyDialoguesCommand(context);
 
 	// Register providers
 	registerCodeLensProvider(context);